api/requirements.txt
fastapi==0.109.1
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.6.1
PyJWT==2.8.0
SQLAlchemy==2.0.27
aiofiles==23.2.1
python-magic==0.4.27
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
nltk==3.8.1
spacy==3.7.4
pdf2image==1.17.0
pytesseract==0.3.10
pillow==10.2.0
pymongo==4.6.1
motor==3.3.2 